requirements.txt
Django>=4.2
djangorestframework
mysqlclient
redis
django-redis
django-elasticsearch-dsl
gunicorn
Faker>=19.0.0